Update npm dependencies and fix corresponding breakages
The "crockford" JSCS preset was updated with a new rule:
"The var statement should be the first statement in the function body."
We don't need this rule, so this rule has been overridden in .jscsrc.

The grunt-contrib-connect plugin now uses connect 3, which moved
middlewares to separate repositories. So now we need to use the serve-static
middleware instead of connect.static.