Update application and test Bower dependencies and fix corresponding breakages
videojs updated its `bower.json`'s `"main"` field to use ES6 and Sass files, which are incompatible
with the way we inject dependencies. So I updated `bower.json` to override that config so that
`grunt bowerInstall` can inject the dependencies correctly.

Gemnasium (https://gemnasium.com/cs-education/sysbuild) reports that modernizr has latest
version 3.2.0, but in fact the latest version is 2.8.3, which we already have. Currently there doesn't
seems to be a way to ignore certain packages in Gemnasium, so we'll have to live with the
 "dependencies out-of-date" badge. This isn't so bad, because we are actually using an outdated
 version of typeahead.js.

typeahead.js cannot be updated to the next major version (0.11.x), because it introduces breaking
changes into its API. It'll take some effort to support the new API. So, we need to stick with 0.10.x
versions. 0.10.5 is the latest version, which we already have.