Open license in a modal window instead of linking to GitHub
LICENSE.md is already bundled with the distribution. So, the license text
is grabbed from the root of the dist/ folder. Because the LICENSE.md is not
in the content root during development, the connect server configuration in
Gruntfile.js had to be modified to handle the /LICENSE.md route.

Closes #123