Moved lessons/chapter index to a new component
Components are created and destroyed every time the "page" (in Router)
is
changed, so the actual chapter content grabbed from sysassets is managed
by
a separate module in lessons.js, so that the state is maintained
(chapters are
not redownloaded, etc.) across instantiations of the lessons-page
component.

I deleted and will continue to delete stuff from app/index.html and
other old
files as I move them into separate components.