Handle delayed chapter loading
Prior to this commit, if someone visited a particular activity directly
(by
directly typing the URL), an error is thrown because the activity is
accessed
before the chapters have been fully loaded.

Now we use Knockout computed observables to observe changes to the
chapters
array, so that the activity view is automatically updated when the
chapters
have been loaded. Computed observables are perfect for these async
scenarios!