Create component for managing the playground layout
A slight change to the UI has also been made: To keep the code modular,
the
playground is treated the same way as other pages in the app, due to
which
the main nav bar is now shared across all "pages", including the
playground
page. The playground currently will still keep its nav bar, although
it'll not
contain the same links as the main nav bar.