Create editor component
An important change is that the original editor module managed user
preferences,  DOM manipulations and updating ko observables in the
sys-view-model module.

The new editor component does not manage any user preference storage,
it is passed knockout observable parameters by the parent component.
These
observables are considered the single source of "truth", and every DOM
update
manipulates these observables only. This way, there is a clean
separation
between the DOM, the ace editor instance, and the parent component
(which
could or could not be managing user preference storage, this component
does
not care).