Remove lesson nav pager & footer from activity-page
The play-activity-page component is a "child" of the activity-page
component, and it included a lesson nav pager component. But
activity-page also contained a lesson nav pager, leading to two
instances of the lesson nav pager component being created for the
play-activity-page. This commit moves the lesson nav pager to the
video-activity-page component, so that it is contained and displayed
by the "child" components of activity-page.

The footer is also moved from the activity-page to the main index.html
page.