Fix editor settings popover
Had to add "e.preventDefault()" to the settings popover click handler,
because clicking on the settings button navigates the browser. This
behavior was not seen in the previous project structure, I'm not sure
why is it occurring with the new project structure. Probable causes
could be a newer version of jQuery or Bootstrap, CSS issues due to
restructuring of the HTML elements, etc. The current fix
(using e.preventDefault()) seems to work fine and free of unintended
side effects.