Migrate editor and tab resizing implementation
Some changes were made to adapt to the new structure. In particular:
1. The man-page-tab CSS class is no longer used, instead, its styles
(full width and height) are hardcoded in all tabs (in editor-pane).
2. The editor resize function no longer accepts the "resizeAfter"
parameter, as it was not used anywhere and it conflicted with
jQuery's event parameter when the function is used with the window
resize event.
3. The original implementation resized the editor and tabs whenever a
tab was selected and shown (by using the "shown.bs.tab" event),
however, this seemed to make no difference and thus redundant in
the new implementation, and so is not used.

There are still some issues with window resizing, scrollbars showing,
etc, but they were also present in the original implementation. So,
I'm not addressing them yet, as the point of the migration is to
get to the same level of functionality as the original implementation.