Migrate core modules
The application is still not functional, because the newly migrated
modules still have to be fixed and wired up correctly.

Mostly of the changes are conversion to ES2015 classes and modules, and
using Arrow functions instead of the more verbose "function ()".

The old router.js is deleted because all of its functionality has
already been migrated.