Fix Jor1k worker path and terminal initialization
The terminal canvas elements had to be moved to index.html, because
Jor1k's terminal expects those elements to be present in the DOM during
initialization. See the comment inside the displayTerminalElements
function in playground-term-pane.js.

Compilation is not functional yet, as some SysViewModel references
need to be updated still.