Remove no-longer-needed overrides in bower.json
Remove the "resolutions" field as well, it is probably not needed.

Also some things I forgot to mention in the last commit message:
* The knockout dependency versions in bower.json was updated by Bower
automatically, probably because there were two entries
* To make Jor1k work with RequireJS, I had to fix an issue with its
use of require: some of Jor1k's modules were requiring
"messagehandler.js" instead of "messagehandler" (without the .js
extension). This caused trouble with RequireJS (which tried to look
for "messagehandler.js.js"), and is not the recommended way to
include modules anyway, even in CommonJS. I had to manually replace
instances of "require('messagehandler.js')" with
"require('messagehandler')", and this will need to be done every time
Jor1k is downloaded. However, I'll soon be opening a pull-request on
the Jor1k project to fix this in the source.