Setup gulp test task, migrate old test
The new project structure uses Jasmine as the testing framework, and
Karma as the test runner. Both of them were configured well by
generator-ko, and I only had to make a few changes to it.

There is also now a Gulp task for running Karma.

The test migrated from the old project, gcc-output-parser.js, doesn't
work (pass) yet because it uses assert() style assertions, as the old
project structure used the chai and mocha libraries. The test needs to
be modified to use assertion styles supported by Jasmine.

Also, testing in the browser without using Karma doesn't work yet,
because we will need to configure a web server that can transpile
ES6 to ES5 on the fly (basically need to configure a "serve:test"
gulp task).