Add support for running tests in browser w/o Karma
generator-ko had already added support for this by including the
test/index.html and SpecRunner.browser.js files. However, a web-server
capable of transpiling ES6 to ES5 still needed to be configured. This
commit makes the following changes to achieve that:

* Create a new gulp task "serve:test" which spins up a web server
allowing a browser to load the test/index.html file correctly.

* To support the new gulp task, the gulp-tasks/babel-transpile helper
module had to be modified to support a different root directory. I
used this as an opportunity to refactor the module - the Babel helper
functions are now wrapped in a class, and a root directory can be
provided during its construction.

* I also added 'test/require.config.js' to the list of files to be
skipped by the Babel Transpiler. The "skip list" could be made
configurable, but it seems overkill at this point.

Apart from this, I also added the 'not-found-page' component to the
list of modules to be included in the build, as I forgot to add it
earlier.