Auto-compile styles when dev server is running
The "serve:dist" gulp task no longer builds the application before
running, because I usually first explicitly run "gulp build" and then
"gulp serve:dist", only to find that the build task is run again,
wasting time. Another reason is that a build must be run when
deploying, but before deploying it is recommended to do run serve:dist,
which again means two builds will be run. Due to this change, the
build + deploy workflow will now look like:
1. gulp build
2. gulp serve:dist (to test out changes)
3. gulp deploy
This is a more efficient and simpler workflow in my opinion.