Use the Sass version of Bootstrap
Note that we now use the Bower package "bootstrap-sass" instead of
"bootstrap-sass-official" from the old project, because the webapp and
gulp-webapp Yeoman generators also made that switch. In any case, both
those Bower packages point to the same (official) GitHub repository
(twbs/bootstrap-sass), whose bower.json lists the package name as
"bootstrap-sass". So, "bootstrap-sass" seems to be the official and
recommended package name to use.