Copy jor1k worker, kernel and fs files during build
The "bower_modules/jor1k/bin/" path is not very pretty, but it is
hard-coded in sys-runtime.js. A "non-hacky" way of changing this path
depending on the environment (development, distribution) would be to
use some sort of configuration file, but including the file and
updating it to match the environment would require non-trivial changes
to the build process, and is thus out-of-scope for this transition.