Fix modal windows and popovers
The modal windows were being shown behind the content and behind the
dark fade background. This is because the modal container elements were
positioned relatively (and some parents were absolutely positioned).
The solution is to dynamically move the element outside of any such
containers, and the safest location to move to is the <body> element.

The gcc options error popover and compiler button tooltips were being
initialized on startup, but because those elements are not present in
the DOM then, the jQuery query returned no elements, and so the target
elements were never initialized. The solution is to initialize the
popover and tooltip when they are guaranteed to be in the DOM, and that
point is the constructor for the Knockout component containing them.

Even after doing this the compile button tooltip was not showing up.
The problem turned out to be that the jQuery UI library also has a
tooltip widget, and this conflicted with the one from Bootstrap. See
http://stackoverflow.com/q/13731400 for solutions to the problem. The
"right" solution needed changes to the load order of libraries and
adding initialization code, which is out-of-scope of the migration. One
of the answers to that Stack Overflow post was to simply use Bootstrap
popovers instead, which is exactly what I did here.