Enhance ESLint config and fix errors reported by it
* Switch to a YAML config file instead of JSON. ESLint support both
".yaml" and ".yml" file extensions, but since we already have a
".travis.yml" file, I decided to use the ".yml" extension.

* Add rules from recommended configuration

* I cut shortcuts for making certain rules happy, including enabling
console logging, having unused variables, allowing the use of the
jQuery "$" variable, and marking Bloodhound as a global (which should
ideally be imported properly by configuring RequireJS).

* Proper solutions to the issues mentioned above, and addition of more
rules including stylistic standards, is out of the scope of this
transition.