Fix Ace dependency loading in distribution build
It took a lot of experimentation and trying out different build
configurations to figure out a clean and scalable way to include the
Ace editor into the project.

A comment on a related GitHub issue mentions: "Using source and then
minifying it with rest of your app might be the simplest method."
(https://github.com/ajaxorg/ace-builds/issues/35#issuecomment-49370051)

Using this advice, I think we finally have a good setup for including
Ace into the project. Here are some notes on the setup:

* We now use the source version of Ace as a dependency, not the built
version.

* The development process is not affected, because our version of
RequireJS handles the loading instead of Ace's loader.

* In fact, our development process is much more simple and efficient,
as developers have granular control over which modules from Ace to
import, and there is no confusion over how modules are loaded
(module paths correspond to the source "lib/ace" directory
structure).

* The required Ace modules need to be explicitly included in the build
process so that they are part of the distribution. Maybe we can
automate discovery of required Ace modules in the future.
Nevertheless, this seems much better than specifying module paths or
including multiple <script> tags (as was done previously).