Run lint before test in ci task, fix lint error
Running the "lint" and "test:headless" tasks in parallel causes the
error "EMFILE: too many open files", probably because Karma loads every
JS file in src/ upon startup. I've seen this error to happen only when
"lint" gets started before "test:headless", and not the other way
around (it could just be coincidence). In any case, this leads to
inconsistent behavior, which is why I did not fix the ci task earlier.

The easiest fix is to run the tasks one after the other. As far as I
know, "gulp.start()" is undocumented and not the "right" way to run
tasks in series. The correct way is probably to define dependencies
correctly, but that would mean editing the "test:headless" task to add
"lint" as a task dependency. This would break the modularity of our
gulp tasks, so I avoided it. In any case, the new version of gulp (4.0)
will bring in a new set of APIs for better task control, running tasks
in series or parallel, and more.