Use source Jor1k worker instead of minified file
The "requirejs-web-workers" plugin for RequireJS is used to "import"
the Jor1k worker. The plugin resolves the given module ID to a URL
using the RequireJS configuration, and then returns an initialized
`Worker` object. This way, there is no need to hardcode paths for web
worker scripts, allowing us to keep all path related configuration in
one place.

Another advantage provided by the plugin is that if `Worker` is not
defined (IE < 10), the plugin will load a fake Worker implementation
so that the rest of the application can utilize the same Worker API
whether the browser supports it or not. I'm not sure if it'll work
with Jor1k, though.

To use RequireJS for loading the Jor1k web worker, a wrapper script is
used which imports and configures RequireJS before importing the
actual worker script. This is only needed during development, as the
build step will bundle RequireJS along with the rest on the worker
into a single script to be used in production.

To support the asynchronous style of module loading used in
AMD/RequireJS, the Jor1k source had to be modified a bit. See
s-macke/jor1k@c605800f71e21e59ff40f27e5de44a564adc89b3.