Remove the "requirejs-web-workers" plugin
The last commit added the "requirejs-web-workers" plugin for RequireJS
to "import" the Jor1k worker. Turns out that the URL to the worker is
not resolved correctly in the built version of the application (built
using the r.js optimizer)*, which forces us to hardcode the path to
the worker script in the import statement. Thus, the main advantage
provided by the plugin of keeping all path related configuration in
one place is not useful to us anymore, and so I'm removing it.

The fake Worker implementation provided by the plugin was a nice
benefit, but not really something we needed --- we can use some other
polyfill if required.

(* The gulp task for building the worker script will be added in a
subsequent commit.)