Restructure require config for worker build
For module ID -> path resolution to work correctly in the jor1k
worker after being built/minified, RequireJS needs the "map" config to
not only be given to the optimizer, but also be made available during
runtime. To do this, the "require.config.js" file is included in the
bundle.

For some reason, this config file is always included after the
RequireJS source, which led to the "require" global exported by
RequireJS to be overwritten by "var require = { /* config */};". The
easiest solution was to rewrite the configuration in the form of
"require.config({ /* config */});", which is what is done in this
commit. Because this new style needs the "require" global to be
defined first, the configuration now needs to be loaded after
RequireJS itself has loaded.