Refactor JS build gulp task for multiple outputs
The refactored code allows easily extending the configuration to
generate more JS files. For example, we currently do not use any
worker tasks in the Ace editor, but in case we need to at some point,
we'd need to build out separate worker scripts just like the Jor1k
worker.

No combination of the "bundles", "modules", or some other RequireJS
config worked properly when I tried. The easiest and cleanest approach
I found is to run the optimizer separately for each output file we
wish to create. This also provides a lot of flexibility, for example,
in choosing exactly how we want our modules to be split up, what needs
to be included in a particular bundle, etc.

The event-stream dependency was updated to support passing an array of
streams to es.concat(), instead of each stream as a separate argument.