Fix the plain-browser unit testing setup
The RequireJS configuration for the test environment had to be
restructured to follow the style used by the runtime configuration.

The old configuration style was to set options on a global "require"
variable, which was then picked up by RequireJS after it has loaded.
The new style is to pass the configuration options to the
"require.config()" method, available after RequireJS has loaded. See
commit 92ea4f8fda7ac8f478cb55b25648a67e8c8ad756 for an explanation
as to why the style of configuration was changed.

The new style also allows cleaner extension/overloading of
configuration options, as opposed to setting/overwriting properties on
a global "require" variable.