Fix Travis CI configuration
In the CI environment:
* Cache the right directory containing Bower dependencies
* We do not use Grunt anymore, so need to install 'grunt-cli'.

Note that there is no need to install 'gulp-cli' globally before
installing NPM dependencies. This is because the CI runner does not
invoke gulp directly, but instead invokes the npm test script using
'npm test', which then invokes the 'gulp ci' command. When using an
npm script, npm first adds the locally installed binaries in
'node_modules' to the PATH, then invokes the script. Because gulp will
already be installed locally (because it is a dependency in
package.json), it is added to the PATH automatically when 'npm test'
is invoked, removing the need for a globally installed gulp.

In a small unrelated change in test/SpecRunner.karma.js, I renamed
requirejs.config(...) to require.config(...), because the 'require'
name has been used throughout the project.