Make Jor1k paths configurable
The hardcoded path strings have been moved into config files. Although
this was not strictly necessary for the Jor1k paths in particular,
having this structure present will let us configure modules cleanly in
the future. For example, the Google Analytics property IDs in the
'ga-tracker' module could probably be moved into these config files.

Currently there are two config files, one used during development and
one in the built distributable app. The RequireJS configuration takes
care of using the right config file. The idea is to provide modules
with the ability to import environment-specific values without needing
to determine the environment.

Longer-term, I want to be able to have separate configuration for dev,
staging, and production environments. This isn't possible to do
cleanly for now, because while the config file is included by
RequireJS in the build step, the exact environment is only known in
the deploy step.