Fix window ReferenceError in jor1k-worker-wrapper
Since the declaration of the window variable in the worker was changed
from a var to a const, the declaration was not hoisted anymore,
causing the window global to not be existent when require.js is loaded
using importScripts. I'm not sure exactly why the error is thrown in
requirejs, because requirejs explicitly checks for an undefined window
using "typeof window". Typing "typeof window" into the Chrome console
in a web worker context throws a ReferenceError, which looks like a
browser bug.