Upgrade to Babel 6
Babel 5 was restructured into separate packages with Babel 6:
* Announcement: https://babeljs.io/blog/2015/10/29/6.0.0
* Setup guide: https://babeljs.io/blog/2015/10/31/setting-up-babel-6
* quick migration guide:
  https://medium.com/@malyw/how-to-update-babel-5-x-6-x-d828c230ec53

We use the recommended "env" preset:
https://babeljs.io/docs/plugins/preset-env/

Babel 6 changed how modules are handled:
http://stackoverflow.com/questions/33505992/babel-6-changes-how-it-exports-default
Hence, some changes to import statements had to be made. I also added
the "add-module-exports" plugin (
https://www.npmjs.com/package/babel-plugin-add-module-exports) to keep
the Babel 5 behavior in Babel 6.

Lastly, the line
    this.localFS = require('fs');
had to be replaced with
    this.localFS = BrowserFS.BFSRequire('fs');
because the Babel AMD module plugin was incorrectly trying to process
the "require" call.